moves inventory to its own file
